move header footer navbar to templates
