+++ index.php
@@ -1,7 +1,7 @@
 <?php require_once "./_autoload.php"; ?>
 
-<?php require_once "./header.php"; ?>
-<?php require_once "./navbar.php"; ?>
+<?php require_once "./view/template/header.php"; ?>
+<?php require_once "./view/template/navbar.php"; ?>
 
 <div class="container">
     <div class="row">
@@ -194,7 +194,7 @@
 
     </div>
 </div>
-<?php require_once "./footer.php"; ?>
+<?php require_once "./view/template/footer.php"; ?>
 <script>
     $(document).ready(function () {
         $('#avbl_devices').DataTable({
