+++ loan.php
@@ -1,6 +1,6 @@
 <?php
-require_once "./header.php";
-require_once "./navbar.php";
+require_once "./view/template/header.php";
+require_once "./view/template/navbar.php";
 ?>
     <div class="container">
         <div class="headline col-12">
@@ -39,4 +39,4 @@ require_once "./navbar.php";
 
         </form>
     </div>
-<?php require_once "./footer.php"; ?>
+<?php require_once "./view/template/footer.php"; ?>
