+++ view/template/navbar.php
